split project view, use a draggable list component,, refactor share app.  !!! order not saved for now !!!
